+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [Unreleased]
+### Added
+- Initial layout for price distribution frontend and related routes [FD-300](https://opensource.ncsa.illinois.edu/jira/browse/FD-300)
+- API call to DW [FD-304](https://opensource.ncsa.illinois.edu/jira/browse/FD-304)
+
 ## [1.3.0] - 2020-06-12
 ### Changed
 - Horizontal user details bar in the Header replaced with 'User Account' Icon with dropdown menu [FD-229](https://opensource.ncsa.illinois.edu/jira/browse/FD-229)